+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - Do not create a generic PackageURL for URLs without a path in url2purl #72
 - Increase the name field size for packageurl.contrib.django.models.PackageURLMixin to 255 #73
 - Use project name as the Package name when creating generic sourceforge PackageURLs #74
+- Update PyPI route pattern in url2purl to handle different file name formats #76
 
 0.9.7 (2022-02-07)
 ------------------